fix warning in VS 2015
git-svn-id: https://svn.code.sf.net/p/xlw/code/branches/Dev6/xlw@1537 948fe115-5b3c-0410-8233-88faca9dced8

+++ xlw/src/XlfExcel.cpp
@@ -101,7 +101,7 @@ namespace
     typedef struct
     {
         HWND hWnd;
-        unsigned short loWord;
+        WORD loWord;
     } GetMainWindowStruct;
 
     BOOL CALLBACK GetMainWindowProc(HWND hWnd, LPARAM lParam)
@@ -138,7 +138,7 @@ HWND xlw::XlfExcel::GetMainWindow()
     XLOPER ret;
     if(Call4(xlGetHwnd, &ret, 0) == xlretSuccess)
     {
-        GetMainWindowStruct getMainWindowStruct = { NULL, ret.val.w};
+        GetMainWindowStruct getMainWindowStruct = { NULL, static_cast<WORD>(ret.val.w)};
         EnumWindows(GetMainWindowProc, (LPARAM) &getMainWindowStruct);
 
         if (getMainWindowStruct.hWnd != NULL)
